pyproject.toml
[project]
name = "hypersquid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hyperliquid-python-sdk>=0.20.0",
    "python-dotenv>=1.0.0",
]
